pyproject.toml
[project]
name = "google-fact-check"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "loguru>=0.7.3",
    "polars>=1.36.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
]
